Bnex-Backend/pyproject.toml
[tool.poetry]
name = "bnex"
version = "0.1.0"
description = "project bnex, technical stage"
authors = ["Juan Carlos Leone Barreto <jc.leone98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
markdown = "^3.6"
django-filter = "^24.3"
pytest = "^8.3.2"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
blue = "^0.9.1"
isort = "^5.13.2"
taskipy = "^1.13.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.5.31"
mkdocstrings = "^0.25.2"
mkdocstrings-python = "^1.10.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue . && isort . "
lint-check = "blue --check --diff . && isort --check --diff . "
docs = "mkdocs serve"
pre_test = "task lint-check && task lint"
test = "docker exec backend python manage.py test"
post_test = "coverage html"
migrate = "docker exec backend python manage.py migrate"
makemigrations = "docker exec backend python manage.py makemigrations"
createsuperuser = "docker exec -it backend python manage.py createsuperuser"
startapp = "docker exec backend python manage.py startapp {param}"
ip-database = "docker inspect -f '{{range.NetworkSettings.Networks}}{{.IPAddress}}{{end}}' postgres-database"


